Send valid request when creating staging repos
Prior to this commit, `Gson` was configured to use a custom
`TypeAdapterFactory` that was supposed to wrap all top-level JSON
objects in another object with a `data` attribute and unwrap received
objects to remove that attribute. However, it was mistakenly used for
nested JSON objects as well. This lead to sending an invalid payload
when creating staging repositories.

Now, a new `Dto` object is used to make wrapping in a `data` attribute
explicit. The `TypeAdapterFactory` is no longer used.

Fixes #30.