Move retrying config from repo to extension and rename
To avoid confusion since publishing is not retried, only checking the
repo's status when transitioning `RetryingConfig` is moved from
`NexusRepository` to `NexusPublishExtension` and renamed to
`TransitionCheckOptions`. The options specified in the extension are by
default applied to all close and release tasks but can be overridden on
each task, if need be.

Resolves #20.