Support publishing using Ivy (#222)
* Spike support for Ivy publications

Ivy publications are used when the layout of the deployed artifacts
needs to be customized, e.g. when publishing Sbt plugins.

* Refactor the nexus publishing plugin to support Ivy

Add a fallback url for the early url check inside IvyPublishingPlugin
Add a the proper task ids for the nexus plugin to depend on for ivy publishing as well
Change publication types to be a single property instead of a set

* Create test for Ivy publishing as well

Move the already existing tests to a BaseNexusPublishPluginTests and move all Ivy/Maven specific test to aptly named test classes

* Add documentation for using with the ivy-publish plugin

* Add a property to allow changing the ivy repository layout

* Add the minimal ivy related e2e tests

* Add the ivy-sbt tests

* Add support for defining the ivy repository pattern layout

* Make maven layout the default for ivy

* Update the README to include ivyPatternLayout

* Revert accidental reformats

* Apply spotless changes

* Update README.md wit suggestions

Co-authored-by: Marc Philipp <marc@gradle.com>

* Make the code compatbile with gradle 5

UrlArtifactRepository was only introduced in gradle 5.1 and gradle modules are not uploaded in gradle 5

* Move publishType and ivyRepositoryPattern to NexusRepository

It makes more sense to have these properties configured on an individual repository level rather than at the plugin level
This will allow the configuration of repositories with different publication types and repository layouts, mimics the original gradle configuration more closely

* Apply code style related review comments from PR #222

* Remove superflous ivy test

* Remove extra whitespace

* Refactor duplicated tests

* Remove gradl5 compatibility code

* Readd the usage of convention for the PublicationType property

* Update the gradle wrapper version to 8.1.1

* Exclude 8.2+ versions of gradle from compat tests, currently in milestone

* Remove references to 5.0 and update docs and version

* Fix the gradle 8.2 incompatibility issue

* Remove methods accidentally added back

* Remove accidentally committed asdf config file

* Add test to verify seting explicit null for publishType and remove nullable parameter

* Move the PublicationType to its correct place and udpate readme

* Fix the kotlin dsl errors

* Fix typo

* Update ivy e2e projects

* Update nexus-publish-e2e-minimal-ivy

Fixed signing.

* Switch e2e modules to nexus-publish group

* Update ivy e2e projects

Old Sonatype Nexus and publicationType place in the extension.

---------

Co-authored-by: Marc Philipp <mail@marcphilipp.de>
Co-authored-by: Marc Philipp <marc@gradle.com>
Co-authored-by: Marcin Zajączkowski <148013+szpak@users.noreply.github.com>