Fix most of configuration cache warnings (#223)
Decouple read of useStaging from the extension by an intermediate property. This helps to remove the captured extension object which results in serializing a lot of unnecessary things.

It prevents:
cannot serialize object of type 'org.gradle.api.DefaultTask', a subtype of 'org.gradle.api.Task', as these are not supported with the configuration cache.