Add manual invalidation of IvyRepository urls
In gradle 8.2 the url became the part of the artifactResourcePatterns are stored with the URL, and the URL is taken from the repository description
RepositoryDescriptions are created when the IvyArtifactRepository is created. The nexus plugin uses a deferred creation of staging repositories and relies on a lazy provider to provide the staging repository url to the artifact repository
This no longer works as the url is no longer read at any later point.

This PR will call the `invalidateDescriptor()` method on the IvyArtifactRepository via reflection if possible.