Allow applying the plugin to a non-root project
This introduces the idea of `independentProjects`, defaulting to false,
and limiting application of the plugin to just the project applied to
if it's true.

By itself, this change should allow projects with a single subproject
(such as is templated by `gradle init`) to apply the plugin to that
project only.  It's also laying a little ground-work for cross-project
synchronisation of staging repositories that's compatible with isolated
projects.